Refactor pseudo-types
