+++ src/PseudoTypes/ArrayKey.php
@@ -11,15 +11,17 @@ declare(strict_types=1);
  * @link      http://phpdoc.org
  */
 
-namespace phpDocumentor\Reflection\Types;
+namespace phpDocumentor\Reflection\PseudoTypes;
 
 use phpDocumentor\Reflection\PseudoType;
 use phpDocumentor\Reflection\Type;
+use phpDocumentor\Reflection\Types\AggregatedType;
+use phpDocumentor\Reflection\Types\Compound;
+use phpDocumentor\Reflection\Types\Integer;
+use phpDocumentor\Reflection\Types\String_;
 
 /**
- * Value Object representing a array-key Type.
- *
- * A array-key Type is the supertype (but not a union) of int and string.
+ * Value Object representing the type `array-key`.
  *
  * @psalm-immutable
  */
