+++ src/PseudoTypes/ClassString.php
@@ -11,17 +11,19 @@ declare(strict_types=1);
  * @link      http://phpdoc.org
  */
 
-namespace phpDocumentor\Reflection\Types;
+namespace phpDocumentor\Reflection\PseudoTypes;
 
 use phpDocumentor\Reflection\Fqsen;
+use phpDocumentor\Reflection\PseudoType;
 use phpDocumentor\Reflection\Type;
+use phpDocumentor\Reflection\Types\String_;
 
 /**
- * Value Object representing the type 'string'.
+ * Value Object representing the type 'class-string'.
  *
  * @psalm-immutable
  */
-final class InterfaceString implements Type
+final class ClassString extends String_ implements PseudoType
 {
     /** @var Fqsen|null */
     private $fqsen;
@@ -32,6 +34,11 @@ final class InterfaceString implements Type
     public function __construct(?Fqsen $fqsen = null)
     {
         $this->fqsen = $fqsen;
+    }
+
+    public function underlyingType(): Type
+    {
+        return new String_();
     }
 
     /**
@@ -48,9 +55,9 @@ final class InterfaceString implements Type
     public function __toString(): string
     {
         if ($this->fqsen === null) {
-            return 'interface-string';
+            return 'class-string';
         }
 
-        return 'interface-string<' . (string) $this->fqsen . '>';
+        return 'class-string<' . (string) $this->fqsen . '>';
     }
 }
