+++ tests/unit/PseudoTypes/InterfaceStringTest.php
@@ -11,13 +11,13 @@ declare(strict_types=1);
  * @link      http://phpdoc.org
  */
 
-namespace phpDocumentor\Reflection\Types;
+namespace phpDocumentor\Reflection\PseudoTypes;
 
 use phpDocumentor\Reflection\Fqsen;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @coversDefaultClass \phpDocumentor\Reflection\Types\InterfaceString
+ * @coversDefaultClass \phpDocumentor\Reflection\PseudoTypes\InterfaceString
  */
 class InterfaceStringTest extends TestCase
 {
